recommender/requirements_pip.txt
--extra-index-url https://download.pytorch.org/whl/cu121
jupyterlab
jupyter
ipykernel
matplotlib
numpy==1.23.4
pandas==2.0.3
scikit_learn==1.3.0
ratelimit
python-dotenv
torch==2.1.0+cu121